% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{owens.T}
\alias{owens.T}
\title{Owen's T function}
\usage{
owens.T(h, a)
}
\arguments{
\item{h}{A numeric vector.}

\item{a}{A numeric vector.}
}
\value{
A numeric vector containing the values of T(h, a).
}
\description{
Computes Owen's T function, T(h, a). This is a user-facing
wrapper for the C++ implementation.
}
\examples{
owens.T(1, 1)
owens.T(c(1, 2, 3), a = 0.5)
}
